[SCSI] SCSI & FC transport: extend event vendor id's to 64bits
During discussions with Mike Christie, I became convinced that we needed
a larger vendor id. This patch extends the id from 32 to 64 bits.

This applies on top of the prior patches that add SCSI transport events
via netlink.

+++ include/scsi/scsi_netlink_fc.h
@@ -59,7 +59,7 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 struct fc_nl_event {
 	struct scsi_nl_hdr snlh;		/* must be 1st element ! */
 	uint64_t seconds;
-	uint32_t vendor_id;
+	uint64_t vendor_id;
 	uint16_t host_no;
 	uint16_t event_datalen;
 	uint32_t event_num;
